requirements.txt
appdirs==1.4.3
cachetools==2.0.1
certifi==2017.11.5
chardet==3.0.4
cryptography==1.7.1
enum34==1.1.6
future==0.16.0
geopy==1.11.0
google-assistant-grpc==0.0.2
google-assistant-library==0.0.3
google-auth==1.2.1
google-auth-oauthlib==0.2.0
google-speech==1.0.16
googleapis-common-protos==1.5.3
gpspy3==0.1.0
grpc-google-cloud-speech-v1beta1==0.14.0
grpcio==1.8.1
httplib2==0.10.3
idna==2.6
iso8601==0.1.12
keyring==10.1
keyrings.alt==1.3
numpy==1.12.1
oauth2client==3.0.0
oauthlib==2.0.6
pbr==3.1.1
PiGPS==0.1.1
protobuf==3.5.0.post1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
pygobject==3.22.0
pyserial==3.4
PySocks==1.5.0
python-apt==1.1.0b5
pyxdg==0.25
PyYAML==3.12
requests==2.18.4
requests-oauthlib==0.8.0
RPi.GPIO==0.6.3
rsa==3.4.2
SecretStorage==2.3.1
serial==0.0.23
six==1.11.0
stevedore==1.28.0
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.8.2
web-cache==1.0.2
